requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
Flask==1.1.2
gunicorn==20.0.4
idna==2.9
idna-ssl==1.1.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
multidict==4.7.6

pkg-resources==0.0.0
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
slackclient==2.6.2
typing-extensions==3.7.4.2
Werkzeug==1.0.1
wincertstore==0.2
yarl==1.4.2
